Fix some issues.
This fixes a few issues that have come about unexpectedly.

First, the POST url for new version has changed, but luckily they
have it in the links section of the target deposit.

Second, the response code for depositing files seems to have changed
from 200 to 201.

Lastly, the metadata publication_date format seems to have changed
to only accept the date, not the time as well.